ENG-3704 Move ApiClaim models into bundle-descriptor
