--- src/models/api.ts
@@ -1,11 +1,0 @@
-export enum ApiType {
-  Internal = 'internal',
-  External = 'external'
-}
-
-export interface ApiClaim {
-  name: string,
-  type: ApiType,
-  serviceId: string,
-  bundleId?: string
-}
