+++ src/services/api-claim-service.ts
@@ -1,5 +1,5 @@
 import { CLIError } from '@oclif/errors'
-import { ApiClaim } from '../models/api'
+import { ApiClaim } from '../models/bundle-descriptor'
 import BundleDescriptorService from './bundle-descriptor-service'
 import { BundleDescriptor } from '../models/bundle-descriptor'
 import { MfeConfigService } from './mfe-config-service'
